Fix : 채팅 상세 페이지- 상세 유저 프로필 모달 외부 아무데나 클릭해도 닫히도록 수정[#71]
