+++ client/src/Template/ChatListTemplate.tsx
@@ -5,6 +5,7 @@ import ChatProfileContainer from "../Organism/ChatProfileContainer";
 import ChatListContainer from "../Organism/ChatListContainer";
 import { getChatsInfo } from "../util/dummyData";
 import { ChatsInfoType } from "../util/type";
+import ProfileModal from "./ProfileModal";
 
 const ChatListTemplateStyle = css`
   width: 80vw;
@@ -19,6 +20,15 @@ const ChatListTemplateStyle = css`
 function ChatListTemplate() {
   const [chatsInfo, setChatsInfo] = useState<ChatsInfoType | null>(null);
   const [clickedRoomIndex, setClickedRoomIndex] = useState(-1);
+  const [openModal, setOpenModal] = useState<number | null>(null);
+  const changeOpenModal = (event: React.MouseEvent) => {
+    const { id } = (event.target as HTMLElement).dataset;
+    if (id === undefined) {
+      setOpenModal(null);
+    } else if (id !== undefined) {
+      setOpenModal(Number(id));
+    }
+  };
   const getChatRoomData = async () => {
     const data = await getChatsInfo();
     setChatsInfo(data);
@@ -28,9 +38,10 @@ function ChatListTemplate() {
   }, []);
 
   return (
-    <div css={ChatListTemplateStyle}>
+    <div css={ChatListTemplateStyle} onClick={changeOpenModal}>
       <ChatProfileContainer chatsInfo={chatsInfo} setClickedRoomIndex={setClickedRoomIndex} />
       <ChatListContainer chatInfo={chatsInfo?.data[clickedRoomIndex]} />
+      {chatsInfo && clickedRoomIndex !== -1 && openModal !== null && <ProfileModal data={chatsInfo.data[clickedRoomIndex].member[openModal]} />}
     </div>
   );
 }
